backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
breeze-connect==1.0.57
PySocks==1.7.1
python-dotenv==1.0.1
httpx==0.27.0
